+++ client/src/components/contacts/display/searchResultsContainer/searchResultsGroup/index.js
